Issue #1404: Moved PreferenceApiService to acl package, removed old tests
