+++ api/src/test/java/com/epam/pipeline/test/web/ControllerTestBeans.java
@@ -48,7 +48,7 @@ import com.epam.pipeline.manager.pipeline.PipelineConfigApiService;
 import com.epam.pipeline.acl.docker.ToolApiService;
 import com.epam.pipeline.manager.pipeline.ToolGroupApiService;
 import com.epam.pipeline.manager.pipeline.ToolManager;
-import com.epam.pipeline.manager.preference.PreferenceApiService;
+import com.epam.pipeline.acl.preference.PreferenceApiService;
 import com.epam.pipeline.manager.preference.PreferenceManager;
 import com.epam.pipeline.acl.region.CloudRegionApiService;
 import com.epam.pipeline.manager.search.SearchManager;
